refactor: move @planx/components types.ts files to model.ts
