+++ editor.planx.uk/src/@planx/components/Checklist/Public.tsx
@@ -1,4 +1,4 @@
-import { Checklist } from "@planx/components/Checklist/types";
+import type { Checklist } from "@planx/components/Checklist/model";
 import Card from "@planx/components/shared/Preview/Card";
 import QuestionHeader from "@planx/components/shared/Preview/QuestionHeader";
 import { useFormik } from "formik";
