+++ editor.planx.uk/src/@planx/components/Content/Editor.tsx
@@ -1,7 +1,12 @@
-import { Content, parseContent } from "@planx/components/Content/types";
+import type { Content } from "@planx/components/Content/model";
+import { parseContent } from "@planx/components/Content/model";
 import { TYPES } from "@planx/components/types";
-import { EditorProps, ICONS } from "@planx/components/ui";
-import { InternalNotes, MoreInformation } from "@planx/components/ui";
+import {
+  EditorProps,
+  ICONS,
+  InternalNotes,
+  MoreInformation,
+} from "@planx/components/ui";
 import { useFormik } from "formik";
 import React from "react";
 import InputRow from "ui/InputRow";
