+++ editor.planx.uk/src/pages/FlowEditor/components/forms/Pay.tsx
@@ -1,8 +1,7 @@
-import { Pay } from "@planx/components/Pay/types";
+import type { Pay } from "@planx/components/Pay/model";
 import { parseMoreInformation } from "@planx/components/shared";
 import { TYPES } from "@planx/components/types";
-import { ICONS } from "@planx/components/ui";
-import { InternalNotes, MoreInformation } from "@planx/components/ui";
+import { ICONS, InternalNotes, MoreInformation } from "@planx/components/ui";
 import { useFormik } from "formik";
 import React from "react";
 import {
